Change User.IdUser type to solve errors

+++ user-registration-api/DefaultDomain/Models/User.cs
@@ -9,7 +9,7 @@ public class User
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     [Key]
     [Column("id")]
-    public int? IdUser { get; set; }
+    public long? IdUser { get; set; }
     
     [MaxLength(short.MaxValue)]
     [Column("username")]
